requirements.txt
# Core AI libraries
torch>=2.0.0
torchvision>=0.15.2
torchaudio>=2.0.1

tensorflow>=2.12.0
tensorflow-hub>=0.14.0

# Image + utilities
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.7.0

# Helpful tools
tqdm>=4.65.0
